Feature/grafana loki (#35)
* init Loki

* add logs push to Loki

* add logs push to Loki

* Feature/compose healthcheck (#34)

* add logs dashboard

* add endpoint for healthcheck

* bugfix DB port

* add data importer container

* remove init data import from sh script

* add DataImporter and Healthchecks. Re-org dependencies

* init Postgresql/Loki/Promtail config

* implement json log with file rotation

* implement json log with file rotation and special init for LOKI log type

* unify postgresql log with Python based containers

* add logging to importer script

* update config file

* update compose file and re-build script to reflect new shared volumes structure

* remove status detail and fund url from response body in fund insertion

* Add echos after each change in config file.

* Add removing backslashes from message to log.

* add logs definition and default timezone

* fix typo

* include grafana logs in scrape config

* separate persistent storage for logs and app data. Add persistent storage for romtail and loki

* add option to re-build without persistend storage clean-up

* update speller dictionary

* replace wget with curl

* add common time zone setting

* adjust rebuild scripts

+++ Re-build_from_scratch.sh
@@ -29,12 +29,16 @@
 # Author:   Stanisław Horna
 # GitHub Repository:  https://github.com/StanislawHornaGitHub/Investment
 # Created:  23-Apr-2024
-# Version:  1.1
+# Version:  1.3
 
 # Date            Who                     What
 # 2024-04-28      Stanisław Horna         Remove downloading quotations and refund calculation,
-#                                         due to implementation of Checker service, 
+#                                         due to implementation of Checker service,
 #                                         which will automatically perform those operations.
+#
+# 2024-05-01      Stanisław Horna         Remove initial data import as it was moved to separate container.
+#
+# 2024-05-04      Stanisław Horna         Add prompt for user about removing persistent data.
 #
 
 # define echo colors
@@ -43,19 +47,18 @@ YELLOW='\033[0;33m'
 RED='\033[0;31m'
 RESET='\033[0m'
 
-Main(){
-
+Main() {
+    askUserAboutStorageRemoval
     checkDockerDaemonRunning
     stopRunningContainers
     removePersistantStorage
     buildCompose
     startCompose
-    waitForPostgreSpinUp
+    #waitForPostgreSpinUp
     showComposeStatus
-    importInitData
 }
 
-checkDockerDaemonRunning(){
+checkDockerDaemonRunning() {
 
     # check if docker is running
     if docker ps -a -q -f name="Certainly Not Exist"; then
@@ -66,40 +69,54 @@ checkDockerDaemonRunning(){
     fi
 }
 
-stopRunningContainers(){
+stopRunningContainers() {
 
     docker compose down
 }
 
-removePersistantStorage(){
+askUserAboutStorageRemoval() {
 
-    printYellowMessage "Removing persistant storage"
-
-    dirPath=$(getDotenvVariable "POSTGRESQL_PERSISTENT_STORAGE_PATH")
-    sudo rm -fr "$dirPath"
+    response=""
+    while [ "$response" != "y" ] && [ "$response" != "n" ]; do
+        echo "Remove persistent data? [y/n]: \c"
+        read -r response
+    done
 }
 
-buildCompose(){
-    
+removePersistantStorage() {
+
+    if [ "$response" = "y" ]; then
+        printYellowMessage "Removing persistant storage"
+
+        dirPath=$(getDotenvVariable "APP_DATA_PATH")
+        sudo rm -fr "$dirPath"
+
+        dirPath=$(getDotenvVariable "APP_LOG_PATH")
+        sudo rm -fr "$dirPath"
+    fi
+}
+
+buildCompose() {
+
     printYellowMessage "Building compose containers"
 
     docker compose build
 }
 
-startCompose(){
+startCompose() {
 
     printYellowMessage "Starting compose containers"
 
     docker compose up -d --force-recreate
 }
 
-waitForPostgreSpinUp(){
+waitForPostgreSpinUp() {
 
     PGcontainerName=$(getDotenvVariable "POSTGRESQL_CONTAINER_NAME")
     waitForContainerInit "$PGcontainerName" "database system is ready to accept connections$" 2 10
 }
 
-showComposeStatus(){
+showComposeStatus() {
 
     docker compose ps
 }
@@ -113,45 +130,28 @@ waitForContainerInit() {
     # init variable
     numOfDBstartups=0
 
-    # loop until SQL engine perform 2 complete startups, 
+    # loop until SQL engine perform 2 complete startups,
     # each startup is announced by log message "database system is ready to accept connections"
-    # after first one Database is initialized from scripts located in "SQL" directory, 
+    # after first one Database is initialized from scripts located in "SQL" directory,
     # once DB init is completed, another restart is performed and container is ready to work.
     while [ "$numOfDBstartups" -lt "$PhraseOccurences" ]; do
-        
+
         sleep 0.1
-        # check if container is still running, 
+        # check if container is still running,
         #   if not display container logs.
         #   if yes invoke DB tests.
 
         # if there were some error during initialization container will not be running
         if [ -z "$(docker ps -q -f name="$DockerContainerName")" ]; then
-        docker logs "$DockerContainerName"
+            docker logs "$DockerContainerName"
 
-        printRedMessage "Container is not running"
-        exit "$ExitCode"
+            printRedMessage "Container is not running"
+            exit "$ExitCode"
         fi
         numOfDBstartups=$(docker logs "$DockerContainerName" 2>&1 | grep -c "$PhraseToLookFor")
     done
 
     printGreenMessage "$DockerContainerName has started"
-}
-
-importInitData(){
-
-    printYellowMessage "Importing initial data"
-
-    dirPath=$(getDotenvVariable "INIT_DATA_IMPORT_DIR_PATH")
-    scriptToRun=$(getDotenvVariable "INIT_DATA_SCRIPT")
-    # Set InitialDataImport directory,
-    # because script is looking for particular JSON files within same directory
-    cd "$dirPath" || exit 2
-
-    # Start script which will invoke appropriate API calls
-    "$scriptToRun"
-
-    # Go back to root directory
-    SetRootDirectory
 }
 
 getDotenvVariable() {
